Move from `file' to `filename' in lots of places.  Did this really make sense?  Hm.  See below:
Well, there is only one place where we use the variable `file', now,
and that spot is documented: the function
`rcp-handle-vc-user-login-name' expects to be called only if `file' is
bound.

Potential breakage alert!

But since there should be a real solution in cooperation with the VC
people, this could be acceptable.