Improve error messages.  (Bug#14808)
* tramp.el (tramp-message-show-progress-reporter-message):
Removed, not needed anymore.
(tramp-error-with-buffer): Show message in minibuffer.  Discard
input before waiting.  Reset connection timestamp.
(with-tramp-progress-reporter): Improve messages.
(tramp-process-actions): Use progress reporter.  Delete process in
case of error.  Improve messages.

* tramp-sh.el (tramp-barf-if-no-shell-prompt): Use
condition-case.  Call `tramp-error-with-buffer' with vector and buffer.
(tramp-maybe-open-connection): The car of
`tramp-current-connection' are the first 3 slots of the vector.